Change duplicate webhook notification behavior.
For both withings and fitbit: mark a notification as processed only if we successfully fetched data.

Sometimes when we fetch data, it's not ready yet. Example fetch response after receiving a sleep notification:
```json
{
  "meta": {
    "retryDuration":3000,
    "state":"pending"
  },
  "sleep":[],
  "summary": {
    "totalMinutesAsleep":0,
    "totalSleepRecords":0,
    "totalTimeInBed":0
  }
}
```
It contains no sleep data.
A duplicate notification was received a few seconds after this one, but we ignored it. Maybe we should have tried the fetch.