Withings: Integrate authlib instead of doing oauth manually.
Configure authlib.

Add compliance hooks for withings behavior which deviates from default behavior expected by authlib.

For the token update hook: we don't have access to the dependency injection in FastAPI, which provides a db session: yet we need to write the new token to the database. Save the db session in a `ContextVar` to make it accessible from the update token hook.

Regarding refreshing the token:
* Remove our custom decorator which was refreshing the token upon a 401 error to a resource request.
* Adapt `test_retry_authentication` to be `test_refresh_token`
* When we get/refresh our access token, save it in the db with an expiration date which is a few minutes sooner than the real one. This will help ensure that we'll fetch a new token when we want to do a resource request and our token is about to expire.