Move fitbit activities from `fitbit_latest_activities` to `fitbit_activities`.
Copy all the fields and the data. Don't copy the unique constraint on user + activity type. We'll now log all the users' activities for each supported type, not just one per user and type.

Add a db migration which:
* creates the new table.
* populates the new table with the data from the old table.
* deletes the old table.

Add a crud function to get the latest activity for a given user and type.
* Use this crud function from the fitbit service, to get the latest activity.
* Use this crud function from tests, to verify that we correctly inserted a new activity.

Add a crud function to get the activity for a user with a given log id.
* Use this from the fitbit service, when checking if an activity exists already.

Remove `relationship`s between `FitbitUser` and `FitbitActivity`, to avoid loading all activities for a user.

Add `created_at` and `updated_at` fields to test fixtures, so we can be sure our pre-existing activity data exists with timestamps in the past. Otherwise, the pre-existing activities will be created with the same timestamp (the resolution is seconds) as the new ones created, and the db query to get the latest activity may not always return the new activities.