Try to reduce some dependencies between modules.
Make withings modules which need oauth settings only depend on a subset of settings, WithingsOAuthSettings, instead of the whole Settings.

For the withings oauth update token callback, remove code regarding the db from the oauth config, and put it in the usecase_update_user_oauth instead.

Move the withings.oauth module to oauth.withingsconfig.

Now the services.withings package only has the api module. Will think of a better place to put it...