Add config for report fields
In the `default_report` block, specify all possible `fields` that can be reported to slack.
Allow overriding these per activity type, in the `report` block of each activity type.
Adapt the use case which posts to slack in realtime, to read this setting and only include the relevant fields in the message.